move `eventHandlers` to common code
